refactor(commonpb): introduce shared yanet-commonpb crate
Each of the 4 CLI crates (plain, vlan, function, pipeline) was
compiling common/commonpb/target.proto independently, producing
incompatible local copies of the same types and forcing trait
impls like FromStr for DevicePipeline to be duplicated.

The recommended tonic approach is to compile shared proto files
once in a dedicated crate and wire consumers via extern_path, so
generated types are canonical and impls live in one place. This
commit introduces common/rust/commonpb (yanet-commonpb) and
migrates all four consumers to use it.
