+++ Cargo.toml
@@ -3,6 +3,7 @@
 resolver = "2"
 
 members = [
+    "common/rust/commonpb",
     "cli/core",
     "cli/modules/common",
     "cli/modules/inspect",
