+++ cli/modules/function/build.rs
@@ -1,16 +1,13 @@
 use core::error::Error;
 
 pub fn main() -> Result<(), Box<dyn Error>> {
-    println!("cargo:rerun-if-changed=../../../common/commonpb/target.proto");
     println!("cargo:rerun-if-changed=../../../controlplane/ynpb/function.proto");
 
     tonic_build::configure()
         .emit_rerun_if_changed(false)
         .build_server(false)
-        .compile_protos(
-            &["common/commonpb/target.proto", "controlplane/ynpb/function.proto"],
-            &["../../.."],
-        )?;
+        .extern_path(".commonpb", "::commonpb::pb")
+        .compile_protos(&["controlplane/ynpb/function.proto"], &["../../.."])?;
 
     Ok(())
 }
