+++ common/rust/commonpb/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "yanet-commonpb"
+version = "0.1.0"
+edition = "2024"
+publish = false
+rust-version = "1.85"
+
+[dependencies]
+prost = "0.13"
+serde = { version = "1", features = ["derive"] }
+tonic = { version = "0.13", features = ["gzip"] }
+
+[build-dependencies]
+tonic-build = "0.13"
